initialisation

+++ src/main/java/ma/emsi/hotelManager/model/Service.java
@@ -25,5 +25,5 @@ public class Service {
 
     private String description;
 
-    private Long prix;
+    private double prix;
 }
